refactor: move the card catalog to a separate component
